# Fixes #771: Correct death row tracking in Trie