--- a/core/storage/src/trie/mod.rs
+++ b/core/storage/src/trie/mod.rs
@@ -170,6 +170,12 @@ impl RcTrieNode {
         Ok(cursor.into_inner())
     }
 
+    fn decode_raw(bytes: &Vec<u8>) -> Result<(Vec<u8>, u32), std::io::Error> {
+        let mut cursor = Cursor::new(&bytes[bytes.len() - 4..]);
+        let rc = cursor.read_u32::<LittleEndian>()?;
+        Ok((bytes[..bytes.len() - 4].to_vec(), rc))
+    }
+
     fn decode(bytes: &Vec<u8>) -> Result<(RawTrieNode, u32), std::io::Error> {
         let node = RawTrieNode::decode(&bytes[..bytes.len() - 4])?;
         let mut cursor = Cursor::new(&bytes[bytes.len() - 4..]);
@@ -192,6 +198,17 @@ impl Trie {
 
     pub fn empty_root() -> CryptoHash {
         CryptoHash::default()
+    }
+
+    fn retrieve_raw_node(&self, hash: &CryptoHash) -> Option<(Vec<u8>, u32)> {
+        if let Ok(Some(bytes)) = self.storage.read().expect(POISONED_LOCK_ERR).get_state(hash) {
+            match RcTrieNode::decode_raw(&bytes) {
+                Ok((bytes, rc)) => Some((bytes, rc)),
+                Err(_) => None,
+            }
+        } else {
+            None
+        }
     }
 
     fn retrieve_node(&self, hash: &CryptoHash) -> Result<TrieNode, String> {
@@ -385,6 +402,9 @@ impl Trie {
         }
     }
 
+    /// Deletes a node from the trie which has key = `partial` given root node.
+    /// Returns (new root node or `None` if this was the node to delete, was it updated).
+    /// While deleting keeps track of all the removed / updated nodes in `death_row`.
     fn delete(
         &self,
         node: TrieNode,
@@ -523,8 +543,12 @@ impl Trie {
         };
         let data = rc_node.encode().expect("Failed to serialize");
         let key = hash(&data);
-        let entry = nodes.entry(key).or_insert((data, 0));
-        entry.1 += 1;
+        if let Some(value) = nodes.get_mut(&key) {
+            value.1 += 1;
+        } else {
+            let node_rc = if let Some((_, rc)) = self.retrieve_raw_node(&key) { rc } else { 0 };
+            nodes.insert(key, (data, node_rc + 1));
+        }
         key
     }
 
@@ -557,14 +581,28 @@ impl Trie {
         let mut db_changes = HashMap::default();
 
         let mut nodes = HashMap::default();
-        // TODO: The reference counting doesn't account for the number of existing nodes in
-        // storage that were not touched by this update.
         let new_root = self.flatten_nodes(root_node, &mut nodes);
-        for (key, (value, rc)) in nodes.drain() {
-            let bytes = RcTrieNode::encode(&value, rc).expect("Failed to serialize");
-            db_changes.insert(key.as_ref().to_vec(), Some(bytes));
+        for (key, (value, mut rc)) in nodes.drain() {
+            if let Some(death_rc) = death_row.get(&key) {
+                rc -= death_rc;
+                death_row.remove(&key);
+            }
+            if rc > 0 {
+                let bytes = RcTrieNode::encode(&value, rc).expect("Failed to serialize");
+                db_changes.insert(key.as_ref().to_vec(), Some(bytes));
+            } else {
+                db_changes.insert(key.as_ref().to_vec(), None);
+            }
         }
-        for (hash, _) in death_row {
+        for (hash, death_rc) in death_row {
+            if let Some((bytes, rc)) = self.retrieve_raw_node(&hash) {
+                if rc - death_rc > 0 {
+                    let bytes =
+                        RcTrieNode::encode(&bytes, rc - death_rc).expect("Failed to serialize");
+                    db_changes.insert(hash.as_ref().to_vec(), Some(bytes));
+                    continue;
+                }
+            }
             db_changes.insert(hash.as_ref().to_vec(), None);
         }
         (db_changes, new_root)
@@ -962,19 +1000,40 @@ mod tests {
     #[test]
     fn test_trie_remove_non_existant_key() {
         let trie = create_trie();
-        let empty_root = Trie::empty_root();
         let mut initial = vec![
             (vec![99, 44, 100, 58, 58, 49], Some(vec![1])),
             (vec![99, 44, 100, 58, 58, 50], Some(vec![1])),
             (vec![99, 44, 100, 58, 58, 50, 51], Some(vec![1])),
         ];
-        let (db_changes, root) = trie.update(&empty_root, initial.drain(..));
+        let (db_changes, root) = trie.update(&Trie::empty_root(), initial.drain(..));
         trie.apply_changes(db_changes).is_ok();
 
         let mut changes = vec![
             (vec![99, 44, 100, 58, 58, 45, 49], None),
             (vec![99, 44, 100, 58, 58, 50, 52], None),
         ];
+        let (db_changes, root) = trie.update(&root, changes.drain(..));
+        trie.apply_changes(db_changes).is_ok();
+        for r in trie.iter(&root).unwrap() {
+            r.unwrap();
+        }
+    }
+
+    #[test]
+    fn test_equal_leafs() {
+        let trie = create_trie();
+        let mut initial = vec![
+            (vec![1, 2, 3], Some(vec![1])),
+            (vec![2, 2, 3], Some(vec![1])),
+            (vec![3, 2, 3], Some(vec![1])),
+        ];
+        let (db_changes, root) = trie.update(&Trie::empty_root(), initial.drain(..));
+        trie.apply_changes(db_changes).is_ok();
+        for r in trie.iter(&root).unwrap() {
+            println!("{:?}", r.unwrap());
+        }
+
+        let mut changes = vec![(vec![1, 2, 3], None)];
         let (db_changes, root) = trie.update(&root, changes.drain(..));
         trie.apply_changes(db_changes).is_ok();
         for r in trie.iter(&root).unwrap() {