--- a/node/configs/src/devnet.rs
+++ b/node/configs/src/devnet.rs
@@ -4,17 +4,26 @@ use std::time::Duration;
 const DEVNET_BLOCK_PERIOD_MS: &str = "100";
 
 pub struct DevNetConfig {
-    /// how often devnet produces blocks
+    /// How often devnet produces blocks.
     pub block_period: Duration,
+    /// Path to storage to replay blocks from.
+    pub replay_storage: String,
 }
 
 pub fn get_args<'a, 'b>() -> Vec<Arg<'a, 'b>> {
-    vec![Arg::with_name("test_block_period")
-        .long("test-block-period")
-        .value_name("TEST_BLOCK_PERIOD")
-        .help("Sets the block production period for devnet, in milliseconds")
-        .takes_value(true)
-        .default_value(DEVNET_BLOCK_PERIOD_MS)]
+    vec![
+        Arg::with_name("test_block_period")
+            .long("test-block-period")
+            .value_name("TEST_BLOCK_PERIOD")
+            .help("Sets the block production period for devnet, in milliseconds")
+            .takes_value(true)
+            .default_value(DEVNET_BLOCK_PERIOD_MS),
+        Arg::with_name("replay_storage")
+            .long("replay-storage")
+            .value_name("REPLAY_PATH")
+            .help("Specify a replay path to re-evaluate blockchain.")
+            .takes_value(true),
+    ]
 }
 
 pub fn from_matches(matches: &ArgMatches) -> DevNetConfig {
@@ -22,5 +31,6 @@ pub fn from_matches(matches: &ArgMatches) -> DevNetConfig {
         .value_of("test_block_period")
         .map(|x| Duration::from_millis(x.parse::<u64>().unwrap()))
         .unwrap();
-    DevNetConfig { block_period }
+    let replay_storage = matches.value_of("replay_storage").unwrap();
+    DevNetConfig { block_period, replay_storage: replay_storage.to_string() }
 }