--- a/node/devnet/src/lib.rs
+++ b/node/devnet/src/lib.rs
@@ -1,17 +1,38 @@
 //! Starts DevNet either from args or the provided configs.
 use std::net::{IpAddr, Ipv4Addr, SocketAddr};
+use std::path::PathBuf;
 use std::sync::Arc;
 use std::time::Duration;
 
 use futures::future;
 use futures::stream::Stream;
 use futures::sync::mpsc::channel;
+use log::info;
 
 use client::Client;
 use configs::{ClientConfig, DevNetConfig, get_devnet_configs, RPCConfig};
 use consensus::passthrough::spawn_consensus;
 use coroutines::client_task::ClientTask;
 use primitives::signer::InMemorySigner;
+use primitives::types::BlockId;
+
+/// Re-applies blocks from the start into new client.
+fn replay_storage(client: Arc<Client>, client_cfg: ClientConfig, other_base_path: &str) {
+    let mut other_client_cfg = client_cfg.clone();
+    other_client_cfg.base_path = PathBuf::from(other_base_path);
+    let other_client = Client::new(&other_client_cfg);
+    info!("Replay storage from {}, last block index = {}", other_base_path, other_client.beacon_client.chain.best_index());
+    let mut index = client.beacon_client.chain.best_index();
+    while index <= other_client.beacon_client.chain.best_index() {
+        let beacon_block =
+            other_client.beacon_client.chain.get_block(&BlockId::Number(index)).unwrap();
+        let shard_block =
+            other_client.shard_client.chain.get_block(&BlockId::Number(index)).unwrap();
+        client.try_import_blocks(beacon_block, shard_block);
+        index += 1;
+    }
+    info!("Finished replaying storage: index={}", client.beacon_client.chain.best_index());
+}
 
 pub fn start() {
     let (client_cfg, devnet_cfg, rpc_cfg) = get_devnet_configs();
@@ -22,6 +43,9 @@ pub fn start_from_configs(client_cfg: ClientConfig, devnet_cfg: DevNetConfig, rp
     let mut client = Client::new(&client_cfg);
     client.signer = Arc::new(InMemorySigner::from_seed("alice.near", "alice.near"));
     let client = Arc::new(client);
+    if !devnet_cfg.replay_storage.is_empty() {
+        replay_storage(client.clone(), client_cfg, &devnet_cfg.replay_storage);
+    }
     start_from_client(client, devnet_cfg, rpc_cfg);
 }
 
@@ -70,7 +94,7 @@ pub fn start_from_client(client: Arc<Client>, devnet_cfg: DevNetConfig, rpc_cfg:
         spawn_consensus(client.clone(), consensus_tx, control_rx, devnet_cfg.block_period);
 
         // Spawn tasks to consume not used channels.
-        tokio::spawn(out_block_rx.for_each(move |_| { future::ok(()) }));
+        tokio::spawn(out_block_rx.for_each(move |_| future::ok(())));
         Ok(())
     });