Added a tool to replay blockchain from another storage:
```
cargo run --package devnet -- --replay-storage=issue771
```