@ilblackdragon 2019-03-26
Also fixes #359