@ilblackdragon 2019-03-26
## Benchmarks

Old:
```
test storage_save_block     ... bench:     235,397 ns/iter (+/- 57,695)
test storage_save_get_block ... bench:       2,050 ns/iter (+/- 254)
test trie_lookup ... bench:     602,977 ns/iter (+/- 67,755)
test trie_update ... bench:     400,528 ns/iter (+/- 41,677)
```

New:
```
running 2 tests
test storage_save_block     ... bench:     238,959 ns/iter (+/- 120,344)
test storage_save_get_block ... bench:       2,079 ns/iter (+/- 549)
test trie_lookup ... bench:     632,620 ns/iter (+/- 67,178)
test trie_update ... bench:     478,058 ns/iter (+/- 64,510)
```

A bit slower update as requires more lookups to keep track of reference counters.